the_shield/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
google-generativeai==0.8.5
httpx==0.28.1
pytest==8.4.1
pypdf==5.1.0
python-multipart==0.0.20
